공지사항수정

+++ src/main/java/com/kh/pp/notice/model/dto/NoticeDto.java
@@ -17,11 +17,11 @@ import lombok.NoArgsConstructor;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class NoticeDto {
 
-	private int noticeNo;
-	private int memberNo;
+	private Long noticeNo;
+	private Long memberNo;
 	private String noticeTitle;
 	private String noticeContent;
-	private int noticeCount;
+	private Integer noticeCount;
 	private Date createDate;
 	private String delYn;
